chore: move static methods to their own modules
