+++ google/cloud/sql/connector/utils.py
@@ -14,6 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
+import concurrent
+from typing import Any
 
 import pymysql.cursors
 from cryptography.hazmat.backends import default_backend
@@ -21,24 +23,20 @@ from cryptography.hazmat.primitives import serialization
 from cryptography.hazmat.primitives.asymmetric import rsa
 
 
-def connect(host, user, password, db_name):
+def immediate_future(object: Any) -> concurrent.futures.Future:
+    """A static method that returns an finished future representing
+    the object passed in.
+
+    :type object: Any
+    :param object: Any object.
+
+    :rtype: concurrent.futures.Future
+    :returns: A concurrent.futures.Future representing the value passed
+        in.
     """
-    Connect method to be used as a custom creator in the SQLAlchemy engine
-    creation.
-    """
-    return pymysql.connect(
-        host=host,
-        user=user,
-        password=password,
-        db=db_name,
-        ssl={
-            "ssl": {
-                "ca": "./ca.pem",
-                "cert": "./cert.pem",
-                "key": "./priv.pem",
-            }  # noqa: E501
-        },
-    )
+    fut: concurrent.futures.Future = concurrent.futures.Future()
+    fut.set_result(object)
+    return fut
 
 
 async def generate_keys():
@@ -76,6 +74,26 @@ async def generate_keys():
     )
 
     return priv_key, pub_key
+
+
+def connect(host, user, password, db_name):
+    """
+    Connect method to be used as a custom creator in the SQLAlchemy engine
+    creation.
+    """
+    return pymysql.connect(
+        host=host,
+        user=user,
+        password=password,
+        db=db_name,
+        ssl={
+            "ssl": {
+                "ca": "./ca.pem",
+                "cert": "./cert.pem",
+                "key": "./priv.pem",
+            }  # noqa: E501
+        },
+    )
 
 
 def write_to_file(serverCaCert, ephemeralCert, priv_key):
